Move stub generator to h2mlstubs
