+++ .gitignore
@@ -9,6 +9,7 @@
 /vscode-debugprotocol
 /superbol-free
 /autofonce
+/h2mlstubs
 /ezlibcob
 *~
 _build
