+++ Makefile
@@ -47,7 +47,7 @@ else
   # check on %context_name in dune files.
   ifeq ($(TARGET_PLAT),linux)
 	${DUNE} build ${DUNE_ARGS} ${DUNE_CROSS_ARGS} @install
-	./scripts/copy-bin.sh superbol-studio-oss superbol-vscode-lib superbol-vscode-platform interop-js-stubs node-js-stubs vscode-js-stubs vscode-languageclient-js-stubs vscode-json vscode-debugadapter vscode-debugprotocol superbol-free superbol_free_lib superbol_preprocs superbol_project superbol_platform cobol_common cobol_parser cobol_ptree ebcdic_lib cobol_lsp ppx_cobcflags pretty cobol_config cobol_indent cobol_indent_old cobol_preproc cobol_data cobol_typeck cobol_unit ez_toml ezr_toml sql_preproc sql_ast sql_parser cobol_cfg autofonce autofonce_core autofonce_lib autofonce_m4 autofonce_share autofonce_patch autofonce_config autofonce_misc ez_win32 ez_call ezlibcob
+	./scripts/copy-bin.sh superbol-studio-oss superbol-vscode-lib superbol-vscode-platform interop-js-stubs node-js-stubs vscode-js-stubs vscode-languageclient-js-stubs vscode-json vscode-debugadapter vscode-debugprotocol superbol-free superbol_free_lib superbol_preprocs superbol_project superbol_platform cobol_common cobol_parser cobol_ptree ebcdic_lib cobol_lsp ppx_cobcflags pretty cobol_config cobol_indent cobol_indent_old cobol_preproc cobol_data cobol_typeck cobol_unit ez_toml ezr_toml sql_preproc sql_ast sql_parser cobol_cfg autofonce autofonce_core autofonce_lib autofonce_m4 autofonce_share autofonce_patch autofonce_config autofonce_misc ez_win32 ez_call h2mlstubs ezlibcob
   endif
 endif
 	./scripts/after.sh build
