+++ drom.toml
@@ -351,4 +351,7 @@ dir = "src/autofonce/ez_win32"
 dir = "src/autofonce/ez_call"
 
 [[package]]
+dir = "src/h2mlstubs"
+
+[[package]]
 dir = "src/tramabol/ezlibcob"
