+++ opam/h2mlstubs.opam
@@ -1,7 +1,7 @@
 # This file was generated by `drom` from `drom.toml`.
 # Do not modify, or add to the `skip` field of `drom.toml`.
 opam-version: "2.0"
-name: "ezlibcob"
+name: "h2mlstubs"
 version: "0.3.1"
 license: "MIT"
 synopsis: "SuperBOL Studio OSS Project"
@@ -49,7 +49,10 @@ install: [
 depends: [
   "ocaml" {>= "4.14.0"}
   "dune" {>= "3.17.0"}
-  "h2mlstubs" {= version}
+  "pretty" {= version}
+  "goblint-cil" {>= "2.0.0"}
+  "ez_file" {>= "0.3.0" & < "1.0.0"}
+  "camlp-streams" {>= "5"}
   "odoc" {with-doc}
   "dune" {>= "3.17.0" & < "3.19.0"}
 ]
