+++ src/h2mlstubs/dune
@@ -1,0 +1,59 @@
+; generated by drom from package skeleton 'driver'
+(executable
+  (name main)
+  (public_name h2mlstubs)
+  (package h2mlstubs)
+  ; use field 'dune-libraries' to add libraries without opam deps
+  (libraries pretty goblint-cil ez_file camlp-streams )
+  (flags (:standard (:include linking.sexp)))
+  ; use field 'dune-stanzas' to add more stanzas here
+  
+  
+  )
+
+; Use `static-clibs` to specify static C libs (without lib prefix)
+; and `static-macos-clibs` and `static-alpine-clibs` for system specific deps
+(rule
+  (targets linking.sexp)
+  (enabled_if (<> %{ocaml-config:system} mingw64))
+  (deps (file linking_flags.sh))
+  (action  (with-stdout-to %{targets}
+    (run bash linking_flags.sh linking.sexp %{env:LINKING_MODE=dynamic} %{ocaml-config:system} ))))
+
+(rule
+  (targets linking.sexp)
+  (enabled_if (= %{ocaml-config:system} mingw64))
+  (deps (file linking_flags.sh))
+  (action  (with-stdout-to %{targets}
+    (run bash -c "echo '()' > linking.sexp"))))
+
+
+(rule
+    (targets version.ml)
+    (deps (:script version.mlt) package.toml)
+    (action (with-stdout-to %{targets} (run %{ocaml} -I +unix unix.cma %{script}))))
+
+(documentation
+  (package h2mlstubs))
+
+; use field 'dune-trailer' to add more stuff here
+(rule
+ (targets common_ml.ml)
+ (deps common/common.ml)
+ (action
+  (with-stdout-to common_ml.ml
+   (progn
+    (echo "let contents = {|")
+    (cat common/common.ml)
+    (echo "|}")))))
+
+(rule
+ (targets common_c.ml)
+ (deps common/common.c)
+ (action
+  (with-stdout-to common_c.ml
+   (progn
+    (echo "let contents = {|")
+    (cat common/common.c)
+    (echo "|}")))))
+
