+++ src/h2mlstubs/generate.ml
@@ -592,6 +592,9 @@ let gen_enum (enum : enum) =
 
 
 
+let add_data buf data =
+  Buffer.add_string buf data
+
 let add_file file buf suffix ext =
   let c_file_base = Filename.chop_suffix file ext in
   let c_file_header = c_file_base ^ suffix ^ ext in
@@ -607,6 +610,9 @@ let generate (contents : contents) ml_file c_file =
 
   let c_buf = Buffer.create 10000 in
   Printf.bprintf c_buf "/* generated by genstubs.ml */\n";
+
+  add_data ml_buf Common_ml.contents;
+  add_data c_buf Common_c.contents;
 
   add_file ml_file ml_buf "_header" ".ml";
   add_file c_file c_buf "_header" ".c";
