+++ src/h2mlstubs/package.toml
@@ -1,0 +1,99 @@
+
+# name of package
+name = "h2mlstubs"
+skeleton = "program"
+
+# version if different from project version
+# version = "0.1.0"
+
+# synopsis if different from project synopsis
+# synopsis = ...
+
+# description if different from project description
+# description = ...
+
+# kind is either "library", "program" or "virtual"
+kind = "program"
+
+# authors if different from project authors
+# authors = [ "Me <me@metoo.org>" ]
+
+# name of a file to generate with the current version
+gen-version = "version.ml"
+
+# supported file generators are "ocamllex", "ocamlyacc" and "menhir" 
+# default is [ "ocamllex", "ocamlyacc" ] 
+# generators = [ "ocamllex", "menhir" ]
+
+# menhir options for the package
+#Example:
+#version = "2.0"
+#parser = { modules = ["parser"]; tokens = "Tokens" }
+#tokens = { modules = ["tokens"]}
+# menhir = ...
+
+# whether all modules should be packed/wrapped (default is true)
+# pack-modules = false
+
+# whether the package can be silently skipped if missing deps (default is false)
+# optional = true
+
+# module name used to pack modules (if pack-modules is true)
+# pack = "Mylib"
+
+# preprocessing options
+#  preprocess = "per-module (((action (run ./toto.sh %{input-file})) mod))" 
+# preprocess = "pps ppx_deriving_encoding"
+
+# files to skip while updating at package level
+skip = ["main.ml"]
+
+# package library dependencies
+#   [dependencies]
+#   ez_file = ">=0.1 <1.3"
+#   base-unix = { libname = "unix", version = ">=base" } 
+[dependencies]
+pretty = "version"
+camlp-streams = "5"
+ez_file = "0.3.0"
+goblint-cil = ">=2.0.0"
+
+# package tools dependencies
+[tools]
+# ...
+
+# package fields (depends on package skeleton)
+#Examples:
+#  dune-stanzas = "(preprocess (pps ppx_deriving_encoding))" 
+#  dune-libraries = "bigstring" 
+#  dune-trailer = "(install (..))" 
+#  opam-trailer = "pin-depends: [..]" 
+#  no-opam-test = "yes" 
+#  no-opam-doc = "yes" 
+#  gen-opam = "some" | "all" 
+#  dune-flags = "(:standard (:include linking.sexp))" 
+#  dune-stanzas = "(foreign_stubs (language c) (names mapfile_stubs))" 
+#  static-clibs = "unix" 
+[fields]
+dune-flags = ":standard (:include linking.sexp)"
+dune-trailer = """
+(rule
+ (targets common_ml.ml)
+ (deps common/common.ml)
+ (action
+  (with-stdout-to common_ml.ml
+   (progn
+    (echo \"let contents = {|\")
+    (cat common/common.ml)
+    (echo \"|}\")))))
+
+(rule
+ (targets common_c.ml)
+ (deps common/common.c)
+ (action
+  (with-stdout-to common_c.ml
+   (progn
+    (echo \"let contents = {|\")
+    (cat common/common.c)
+    (echo \"|}\")))))
+"""
