+++ src/tramabol/ezlibcob/libcob_stubs_trailer.c
@@ -1,1 +1,13 @@
 #2 "src/ezlibcob/libcob_stubs_trailer.c"
+/**************************************************************************/
+/*                                                                        */
+/*                        SuperBOL OSS Studio                             */
+/*                                                                        */
+/*  Copyright (c) 2022-2026 OCamlPro SAS                                  */
+/*                                                                        */
+/* All rights reserved.                                                   */
+/* This source code is licensed under the GNU Affero General Public       */
+/* License version 3 found in the LICENSE.md file in the root directory   */
+/* of this source tree.                                                   */
+/*                                                                        */
+/**************************************************************************/
